Initial convert to v3
With Craft v3, some of the syntax has changed and conventions need to be
updated.

Using https://pluginfactory.io/, I generated a new plugin and have
migrated the content accross
