+++ .gitignore
@@ -1,4 +1,19 @@
-*.idea/*
-*.log
-*.DS_Store
-*Thumbs.db
+# COMPOSER
+/vendor
+
+# BUILD FILES
+/bower_components/*
+/node_modules/*
+/build/*
+
+# MISC FILES
+.cache
+.DS_Store
+.idea
+.project
+.settings
+*.esproj
+*.sublime-workspace
+*.sublime-project
+*.tmproj
+*.tmproject
