+++ CHANGELOG.md
@@ -1,0 +1,9 @@
+# Anchors Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## 1.3.1 - 2017.03.16
+### Added
+- Initial release
