+++ README.md
@@ -1,79 +1,37 @@
-# Anchors plugin for Craft
+# Anchors plugin for Craft CMS 3.x
 
-This plugin makes it possible to automatically add linkable anchors to HTML headings in Craft.
+Add linkable anchors to HTML headings in Craft.
 
-The anchors are named based on the heading text. The algorithm Anchors uses to convert the heading text to IDs is similar to Craft’s algorithm for automatically generating entry slugs.
+![Screenshot](resources/img/plugin-logo.png)
 
 ## Installation
 
-To install Anchors, copy the anchors/ folder into craft/plugins/, and then go to Settings > Plugins and click the “Install” button next to “Anchors”.
+To install Anchors, follow these steps:
 
-## Templating
+1. Download & unzip the file and place the `anchors` directory into your `craft/plugins` directory
+2.  -OR- do a `git clone https://github.com/mikestreety/anchors.git` directly into your `craft/plugins` folder.  You can then update it with `git pull`
+3.  -OR- install with Composer via `composer require mikestreety/anchors`
+4. Install plugin in the Craft Control Panel under Settings > Plugins
+5. The plugin folder should be named `anchors` for Craft to see it.  GitHub recently started appending `-master` (the branch name) to the name of the folder for zip file downloads.
 
-To use Anchors in your templates, just pass some HTML into the `|anchors` filter.
+Anchors works on Craft 3.x.
 
-```jinja
-{{ entry.body|anchors }}
-```
+## Anchors Overview
 
-By default, the `anchors` filter will only search for `<h1>`, `<h2>`, and `<h3>` tags. You can customize which tags it searches for by passing in a comma-separated list of tag names.
+-Insert text here-
 
-```jinja
-{{ entry.body|anchors('h2,h3') }}
-```
+## Configuring Anchors
 
-## Configuration
+-Insert text here-
 
-To configure Anchors, create a new “anchors.php” file within the craft/config folder, which returns an array.
+## Using Anchors
 
-The following config settings are supported:
+-Insert text here-
 
-- **anchorClass** – The class name that should be given to named anchors. (Default is `null`, meaning no class will be given.)
-- **anchorLinkClass** – The class name that should be given to anchor links. (Default is `'anchor'`.)
-- **anchorLinkText** – The visible text that anchor links should have. (Default is `'#'`'.)
-- **anchorLinkTitleText** – The title/alt text that anchor links should have. If `{heading}` is included, it will be replaced with the heading text the link is associated with. (Default is `'Direct link to {heading}'`.)
+## Anchors Roadmap
 
-## Plugin API
+Some things to do, and ideas for potential features:
 
-Other plugins can take advantage of Anchors using the provided API.
+* Release it
 
-```php
-$parsedHtml = craft()->anchors->parseHtml($html);
-```
-
-Like the `|anchors` templating filter, `parseHtml()` also allows you to specify which HTML tags should get anchors.
-
-```php
-$parsedHtml = craft()->anchors->parseHtml($html, 'h2,h3');
-```
-
-You can also pass some heading text directly into Anchors to get its generated anchor name:
-
-```php
-$anchorName = craft()->anchors->generateAnchorName($headingText);
-```
-
-## Changelog
-
-### 1.3.1
-
-* Fixed a bug where non-breaking spaces within headings could result in named anchors with “nbsp” in the name.
-
-### 1.3
-
-* Added ‘anchorClass’, ‘anchorLinkClass’, ‘anchorLinkText’, and ‘anchorLinkTitleText’ config settings.
-* The plugin now creates separate named anchor elements that are placed before the headings, rather than adding an `id` attribute to the headings.
-
-### 1.2
-
-* Updated to take advantage of new Craft 2.5 plugin features.
-
-### 1.1
-
-* Added AnchorsService, available globally from `craft()->anchors`, with two public methods:
-  * `parseHtml()`
-  * `generateAnchorName()`
-
-### 1.0
-
-* Initial release
+Brought to you by [Mike Street](https://www.mikestreety.co.uk)
