--- anchors/config.php
@@ -1,8 +1,0 @@
-<?php
-
-return array(
-	'anchorClass' => null,
-	'anchorLinkClass' => 'anchor',
-	'anchorLinkText' => '#',
-	'anchorLinkTitleText' => 'Direct link to {heading}',
-);
