+++ src/translations/en/anchors.php
@@ -1,0 +1,25 @@
+<?php
+/**
+ * Anchors plugin for Craft CMS 3.x
+ *
+ * Add linkable anchors to HTML headings in Craft.
+ *
+ * @link      https://www.mikestreety.co.uk
+ * @copyright Copyright (c) 2017 Mike Street
+ */
+
+/**
+ * Anchors en Translation
+ *
+ * Returns an array with the string to be translated (as passed to `Craft::t()`) as
+ * the key, and the translation as the value.
+ *
+ * http://www.yiiframework.com/doc-2.0/guide-tutorial-i18n.html
+ *
+ * @author    Mike Street
+ * @package   Anchors
+ * @since     1.3.1
+ */
+return [
+    'plugin loaded' => 'plugin loaded',
+];
